+++ src/yajl_encode.h
@@ -39,12 +39,12 @@
 void yajl_string_encode2(const yajl_print_t printer,
                          void * ctx,
                          const unsigned char * str,
-                         unsigned int length);
+                         size_t length);
 
 void yajl_string_encode(yajl_buf buf, const unsigned char * str,
-                        unsigned int length);
+                        size_t length);
 
 void yajl_string_decode(yajl_buf buf, const unsigned char * str,
-                        unsigned int length);
+                        size_t length);
 
 #endif
